requirements.txt
# Dependências principais
py4j==0.10.9.7
pyspark==3.5.5
delta-spark==3.0.0
dbt-core==1.7.0
dbt-spark==1.7.0
dbt-postgres==1.7.0

# Dependências utilitárias
PyYAML==6.0
pandas==1.5.3
numpy==1.24.3

# Dependências de desenvolvimento
pytest==7.3.1
black==23.3.0
flake8==6.0.0